requirements.txt
fastapi
uvicorn[standard]
databases
psycopg2-binary
langchain
pydantic
aiofiles
aiosqlite
python-multipart
requests
python-dotenv
sqlalchemy[asyncio]
langchain_groq
sqlalchemy
duckduckgo-search
langchain-community